db definitions splitted into files
