+++ gql_events/DBDefinitions/EventCategoryModel.py
@@ -1,0 +1,18 @@
+from sqlalchemy import Column, DateTime, String, ForeignKey
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql.functions import now
+
+from .uuid import UUIDColumn, UUIDFKey
+from .base import BaseModel
+
+
+class EventCategoryModel(BaseModel):
+    __tablename__ = "eventcategories"
+    id = UUIDColumn()
+    name = Column(String)
+    name_en = Column(String)
+
+    created = Column(DateTime, server_default=now())
+    lastchange = Column(DateTime, server_default=now())
+    createdby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
+    changedby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
