+++ gql_events/DBDefinitions/__init__.py
@@ -1,3 +1,60 @@
 from .base import BaseModel
+
+from .EventCategoryModel import EventCategoryModel
+from .EventGroupModel import EventGroupModel
 from .EventModel import EventModel
+from .EventTypeModel import EventTypeModel
+
+from .InvitationTypeModel import InvitationTypeModel
+
+from .PresenceModel import PresenceModel
+from .PresenceTypeModel import PresenceTypeModel
+
 from .uuid import UUIDColumn, UUIDFKey
+
+
+
+from sqlalchemy.ext.asyncio import create_async_engine
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.orm import sessionmaker
+from sqlalchemy.exc import NoReferencedTableError
+
+async def startEngine(connectionstring, makeDrop=False, makeUp=True):
+    """Provede nezbytne ukony a vrati asynchronni SessionMaker"""
+    asyncEngine = create_async_engine(connectionstring)
+
+    async with asyncEngine.begin() as conn:
+        if makeDrop:
+            await conn.run_sync(BaseModel.metadata.drop_all)
+            print("BaseModel.metadata.drop_all finished")
+        if makeUp:
+            try:
+                await conn.run_sync(BaseModel.metadata.create_all)
+                print("BaseModel.metadata.create_all finished")
+            except NoReferencedTableError as e:
+                print(e)
+                print("Unable automaticaly create tables")
+                return None
+
+    async_sessionMaker = sessionmaker(
+        asyncEngine, expire_on_commit=False, class_=AsyncSession
+    )
+    return async_sessionMaker
+
+
+import os
+
+
+def ComposeConnectionString():
+    """Odvozuje connectionString z promennych prostredi (nebo z Docker Envs, coz je fakticky totez).
+    Lze predelat na napr. konfiguracni file.
+    """
+    user = os.environ.get("POSTGRES_USER", "postgres")
+    password = os.environ.get("POSTGRES_PASSWORD", "example")
+    database = os.environ.get("POSTGRES_DB", "data")
+    hostWithPort = os.environ.get("POSTGRES_HOST", "localhost:5432")
+
+    driver = "postgresql+asyncpg"  # "postgresql+psycopg2"
+    connectionstring = f"{driver}://{user}:{password}@{hostWithPort}/{database}"
+
+    return connectionstring
