--- gql_events/DBDefinitionsOld.py
@@ -1,188 +1,0 @@
-from email.policy import default
-import sqlalchemy
-import datetime
-import uuid
-
-from sqlalchemy import (
-    Column,
-    String,
-    BigInteger,
-    Integer,
-    DateTime,
-    ForeignKey,
-    Sequence,
-    Table,
-    Boolean,
-    Uuid
-)
-from sqlalchemy.dialects.postgresql import UUID
-from sqlalchemy.orm import relationship
-from sqlalchemy.ext.declarative import declarative_base
-
-from .DBDefinitions import UUIDColumn, UUIDFKey
-from .DBDefinitions import BaseModel
-
-def newUuidAsString():
-    return f"{uuid.uuid1()}"
-
-
-
-# id = Column(UUID(as_uuid=True), primary_key=True, server_default=sqlalchemy.text("uuid_generate_v4()"),)
-
-###########################################################################################################################
-#
-###########################################################################################################################
-
-
-###########################################################################################################################
-#
-# zde definujte sve SQLAlchemy modely
-# je-li treba, muzete definovat modely obsahujici jen id polozku, na ktere se budete odkazovat
-#
-###########################################################################################################################
-
-
-class EventTypeModel(BaseModel):
-    __tablename__ = "eventtypes"
-
-    id = UUIDColumn()
-    name = Column(String)
-    name_en = Column(String)
-    category_id = Column(ForeignKey("eventcategories.id"), index=True, nullable=True)
-    created = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    lastchange = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    createdby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-    changedby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-
-    events = relationship("EventModel", back_populates="eventtype")
-    
-
-class EventGroupModel(BaseModel):
-    __tablename__ = "events_groups"
-    id = UUIDColumn()
-    event_id = Column(ForeignKey("events.id"), index=True)
-    group_id = UUIDFKey()#Column(ForeignKey("groups.id"), index=True)
-                #Column(ForeignKey("groups.id"), index=True)
-                #UUIDFKey()
-
-    created = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    lastchange = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    createdby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-    changedby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-
-    event = relationship("EventModel")
-    #group = relationship("GroupModel")
-
-class EventCategoryModel(BaseModel):
-    __tablename__ = "eventcategories"
-    id = UUIDColumn()
-    name = Column(String)
-    name_en = Column(String)
-
-    created = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    lastchange = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    createdby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-    changedby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-
-    
-
-
-
-##########################################################
-#
-# Zmena
-#
-##########################################################
-class PresenceModel(BaseModel):
-    __tablename__ = "events_users"
-    id = UUIDColumn()
-
-    event_id = Column(ForeignKey("events.id"), index=True)
-    user_id = UUIDFKey()#Column(ForeignKey("users.id"), index=True)
-    
-    invitation_id = Column(ForeignKey("eventinvitationtypes.id"), index=True)
-    presencetype_id = Column(ForeignKey("eventpresencetypes.id"), index=True, nullable=True)
-
-    created = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    lastchange = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    createdby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-    changedby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-
-class PresenceTypeModel(BaseModel):
-    __tablename__ = "eventpresencetypes"
-    id = UUIDColumn()
-
-    name = Column(String)
-    name_en = Column(String)
-    # present, vacantion, ...
-
-    created = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    lastchange = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    createdby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-    changedby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-
-class InvitationTypeModel(BaseModel):
-    __tablename__ = "eventinvitationtypes"
-    id = UUIDColumn()
-
-    name = Column(String)
-    name_en = Column(String)
-    # initiator, invited mandatory, invited voluntary, accepted, tentatively accepted, rejected,
-
-    created = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    lastchange = Column(DateTime, server_default=sqlalchemy.sql.func.now())
-    createdby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-    changedby = UUIDFKey(nullable=True)#Column(ForeignKey("users.id"), index=True, nullable=True)
-
-##########################################################
-#
-# Zmena konec
-#
-##########################################################
-
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-
-from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.ext.asyncio import create_async_engine
-
-
-async def startEngine(connectionstring, makeDrop=False, makeUp=True):
-    """Provede nezbytne ukony a vrati asynchronni SessionMaker"""
-    asyncEngine = create_async_engine(connectionstring)
-
-    async with asyncEngine.begin() as conn:
-        if makeDrop:
-            await conn.run_sync(BaseModel.metadata.drop_all)
-            print("BaseModel.metadata.drop_all finished")
-        if makeUp:
-            try:
-                await conn.run_sync(BaseModel.metadata.create_all)
-                print("BaseModel.metadata.create_all finished")
-            except sqlalchemy.exc.NoReferencedTableError as e:
-                print(e)
-                print("Unable automaticaly create tables")
-                return None
-
-    async_sessionMaker = sessionmaker(
-        asyncEngine, expire_on_commit=False, class_=AsyncSession
-    )
-    return async_sessionMaker
-
-
-import os
-
-
-def ComposeConnectionString():
-    """Odvozuje connectionString z promennych prostredi (nebo z Docker Envs, coz je fakticky totez).
-    Lze predelat na napr. konfiguracni file.
-    """
-    user = os.environ.get("POSTGRES_USER", "postgres")
-    password = os.environ.get("POSTGRES_PASSWORD", "example")
-    database = os.environ.get("POSTGRES_DB", "data")
-    hostWithPort = os.environ.get("POSTGRES_HOST", "localhost:5432")
-
-    driver = "postgresql+asyncpg"  # "postgresql+psycopg2"
-    connectionstring = f"{driver}://{user}:{password}@{hostWithPort}/{database}"
-
-    return connectionstring
